# core-banking-service/docs/swagger.yaml
definitions:
  customer.CreateCustomerRequest:
    properties:
      customer_type:
        $ref: '#/definitions/customer.CustomerType'
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
    type: object
  customer.Customer:
    properties:
      createdAt:
        type: string
      customerNumber:
        type: string
      customerType:
        $ref: '#/definitions/customer.CustomerType'
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      phoneNumber:
        type: string
      status:
        $ref: '#/definitions/customer.CustomerStatus'
      trash:
        type: boolean
      trashedAt:
        type: string
      updatedAt:
        type: string
    type: object
  customer.CustomerStatus:
    enum:
    - ACTIVE
    - CLOSED
    - DORMANT
    - FROZEN
    - DECEASED
    type: string
    x-enum-varnames:
    - CustomerStatusActive
    - CustomerStatusClosed
    - CustomerStatusDormant
    - CustomerStatusFrozen
    - CustomerStatusDeceased
  customer.CustomerType:
    enum:
    - INDIVIDUAL
    - BUSINESS
    type: string
    x-enum-varnames:
    - CustomerTypeIndividual
    - CustomerTypeBusiness
info:
  contact: {}
paths:
  /customers:
    get:
      description: Get all customers.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of customers per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/customer.Customer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Creates a new customer.
      parameters:
      - description: Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.Customer'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a customer
      tags:
      - customers
  /customers/{customerNumber}:
    delete:
      description: Delete Customer by customer number
      parameters:
      - description: Customer number
        in: path
        name: customerNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete Customer by customer number
      tags:
      - customers
swagger: "2.0"